refactor(components): dissolve the components/modals/ container
