+++ src/views/audio-reader/upload-lesson-modal/script-select.vue
